pyproject.toml
[project]
name = "autoresearch"
version = "0.1.0"
description = "Autonomous multimodal embedding model research"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=2.18.0",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "Pillow>=10.2.0",
    "pyarrow>=21.0.0",
    "requests>=2.32.0",
    "torch==2.9.1",
    "torchvision",
    "transformers>=4.48.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128" },
]
torchvision = [
    { index = "pytorch-cu128" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
